change type AssetId = u64

+++ runtime/phoenix/src/lib.rs
@@ -979,7 +979,7 @@ parameter_types! {
 }
 
 parameter_type_with_key! {
-    pub ExistentialDeposits: |_asset_id: u32| -> Balance {
+    pub ExistentialDeposits: |_asset_id: u64| -> Balance {
         Zero::zero()
     };
 }
@@ -1000,7 +1000,8 @@ impl bholdus_tokens::Config for Runtime {
     type Event = Event;
     type Balance = Balance;
     type Amount = Amount;
-    type AssetId = u32;
+    type AssetId = u64;
+
     type Currency = Balances;
     type BasicDeposit = BasicDeposit;
     type FieldDeposit = FieldDeposit;
